Parse d MMMM DateTime inputs

+++ payslip/payslip/Employee.cs
@@ -3,7 +3,7 @@ namespace payslip
     public class Employee
     {
 
-        public Employee(string firstName, string lastName, float salary, float superRate)
+        public Employee(string firstName, string lastName, double salary, double superRate)
         {
             FirstName = firstName;
             LastName = lastName;
@@ -12,8 +12,8 @@ namespace payslip
         }
         public string FirstName { get; private set; }
         public string LastName { get; private set; }
-        public float Salary { get; private set; }
-        public float SuperRate { get; private set; }
+        public double Salary { get; private set; }
+        public double SuperRate { get; private set; }
         
         
     }
